add a line graph test to WCC
